Move infrared entity classes from __init__.py to entity.py
Move EntityDescription and Entity derived classes to a dedicated
entity.py module to satisfy lint rule C7411. All classes are
re-exported from __init__.py to maintain backward compatibility.
